backend/pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"


[project]
name = "EKZ"
version = "0.1.0"
requires-python = ">=3.11"

dependencies = [
    "alembic==1.13.2",
    "bcrypt>=4.3.0",
    "beautifulsoup4>=4.13.4",
    "cryptography>=45.0.5",
    "env>=0.1.0",
    "fastapi==0.115.12",
    "isbnlib>=3.10.14",
    "pika>=1.3.2",
    "psycopg2-binary==2.9.10",
    "pydantic==2.10.6",
    "PyJWT==2.10.1",
    "pymongo==4.5.0",
    "python-dotenv==1.0.1",
    "python-multipart>=0.0.20",
    "requests==2.32.3",
    "ruff>=0.11.2",
    "sqlalchemy==2.0.38",
    "timedelta>=2020.12.3",
    "uvicorn==0.34.0",
]

[tool.setuptools]
packages = ["errors", "validators", "models", "enums"]

[project.optional-dependencies]
dev = ["ruff==0.11.2"]


[tool.ruff]
line-length = 120

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".ipynb_checkpoints",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    ".vscode",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "site-packages",
    "venv",
]

target-version = "py313"


[tool.ruff.lint]

select = ["ALL"]
fixable = ["ALL"]
unfixable = []


ignore = [
    "Q000",
    "Q003",
    #### modules
    "ANN", # flake8-annotations
    "COM", # flake8-commas
    "C90", # mccabe complexity
    "DJ",  # django
    "EXE", # flake8-executable
    "T10", # debugger
    "TID", # flake8-tidy-imports
    #### specific rules
    # "D100",   # ignore missing docs
    "D203",
    # "D101",
    # "D102",
    # "D103",
    # "D104",
    # "D105",
    # "D106",
    # "D107",
    # "D200",
    # "D205",
    "D212",
    # "D400",
    # "D401",
    # "D415",
    # "E402",   # false positives for local imports
    # "E501",   # line too long
    # "TRY003", # external messages in exceptions are too verbose
    # "TD002",
    # "TD003",
    # "FIX002", # too verbose descriptions of todos
    "COM812",

]


[tool.ruff.format]
quote-style = "single"

indent-style = "space"

line-ending = "auto"
